requirements.txt
# Core dependencies - All versions tested and compatible
python-telegram-bot==13.15
solana==0.27.2
solders==0.10.1
aiohttp==3.8.4
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
python-dotenv==1.0.0
cachetools==5.3.0
base58==2.1.1
requests==2.31.0
flask==2.2.5
gunicorn==20.1.0
async-timeout==4.0.2